silence (most but not all) warnings on implicit type conversions from larger integers to smaller ones when compiling with -Wconversion
git-svn-id: https://svn.mcs.anl.gov/repos/parallel-netcdf/trunk@2021 c368ceb8-fd28-0410-a457-e46ee1d7cd4a

+++ src/lib/utf8proc.c
@@ -535,7 +535,7 @@ ssize_t utf8proc_map(
   }
   {
     int32_t *newptr;
-    newptr = (int32_t*) realloc(buffer, result+1);
+    newptr = (int32_t*) realloc(buffer, (size_t)result+1);
     if (newptr) buffer = newptr;
   }
   *dstptr = (uint8_t *)buffer;
